SOLR-12607: Fixes testSplitWithChaosMonkey to incorporate changes made by SOLR-11665.
Artifacts from a failed shard split are aggressively cleaned up by SOLR-11665 so we cannot find or test for any leftover state. This commit removes the check which tried to distinguish between a failure due to failed replica creation and a failure due to an inability to perform soft commit because in both cases, state is cleaned up already. All we assert is that if the split is successful then we eventually switch slice state. In any case, we continue to assert that all shards are consistent and no data is lost. This commit also starts logging DEBUG messages from OverseerTaskProcessor.