SOLR-12607: Set the correct sub-shard name for StdNode instead of parent shard name
The DistributedUpdateProcessor's doFinish() method relies on the failing node's shard name to decide whether to return an exception to the client or to continue. During shard split, we forward updates synchronously to sub shard leaders and we'd like to return exceptions to clients so that they can retry. However, due to using the parent shard name in the StdNode, the comparison between the current replica's shard and StdNode's shard always passed and the errors were never returned.