SOLR-12803: Ensure CUSC routes docs to right cores
ConcurrentUpdateSolrClient can batch together many documents when making
an indexing request to Solr.  When adding an update request to the
current batch being made, it checks that the query-parameters of the
docs being added match those already in the batch.  But prior to this
commit it never checked that the collections/cores were the same.

This could result in documents being sent to the wrong collection if the
same client is used to index documents to two different
cores/collections simultaneously.

This commit addresses this problem, ensuring that documents aren't added
to a batch directed at a different core/collection.